# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EZ Remote Agent",
    "description": "Installs, signs in and runs coding agents.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/agents": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "List agents",
        "description": "Returns every agent with its install and sign-in state.",
        "operationId": "listAgents",
        "responses": {
          "200": {
            "description": "Agents",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AgentStatus"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not signed in to the manager",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/agents/{agent}/install": {
      "post": {
        "tags": [
          "agents"
        ],
        "summary": "Install or update an agent",
        "description": "Installs the newest release on the agent's release channel when it is newer than the installed one, and marks the agent configured.",
        "operationId": "installAgent",
        "parameters": [
          {
            "name": "agent",
            "in": "path",
            "description": "The agent",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Agent"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Installed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentStatus"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in to the manager",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "502": {
            "description": "The download or checksum failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/folders": {
      "get": {
        "tags": [
          "folders"
        ],
        "summary": "List the folders in /projects",
        "operationId": "listFolders",
        "responses": {
          "200": {
            "description": "Folders sorted by name",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FolderStatus"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not signed in to the manager",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/folders/{name}/remote-control": {
      "put": {
        "tags": [
          "folders"
        ],
        "summary": "Choose whether the Claude app lists a folder",
        "description": "Starts or stops the folder's Remote Control server.",
        "operationId": "chooseToServeFolder",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "The folder's name in /projects",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServeBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Saved"
          },
          "401": {
            "description": "Not signed in to the manager",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "No such folder",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Agent": {
        "type": "string",
        "description": "A command-line coding agent the manager can install and run.",
        "enum": [
          "claude",
          "codex"
        ]
      },
      "AgentStatus": {
        "type": "object",
        "description": "One agent's install and sign-in state.",
        "required": [
          "agent",
          "configured",
          "logged_in"
        ],
        "properties": {
          "account": {
            "type": [
              "string",
              "null"
            ],
            "description": "The signed-in account as the CLI reports it, absent when unknown."
          },
          "agent": {
            "$ref": "#/components/schemas/Agent"
          },
          "available_update": {
            "type": [
              "string",
              "null"
            ],
            "description": "A newer release than the installed version, absent when none is known."
          },
          "config_disk_bytes": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Bytes the agent keeps on /config, absent when they cannot be measured.",
            "minimum": 0
          },
          "configured": {
            "type": "boolean",
            "description": "Installed through the manager at least once."
          },
          "install_progress": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/DownloadProgress",
                "description": "Present while an install or update is downloading."
              },
              {
                "type": "null"
              }
            ]
          },
          "installed_version": {
            "type": [
              "string",
              "null"
            ],
            "description": "Absent when the agent is not installed."
          },
          "logged_in": {
            "type": "boolean"
          },
          "login_prompt": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/LoginPrompt",
                "description": "Present while a sign-in is waiting for the person signing in."
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
...
      "DownloadProgress": {
        "type": "object",
        "description": "How far a download has got.",
        "required": [
          "received_bytes"
        ],
        "properties": {
          "received_bytes": {
            "type": "integer",
            "format": "int64",
            "description": "Bytes received so far.",
            "minimum": 0
          },
          "total_bytes": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Size of the download in bytes, absent when the server does not say.",
            "minimum": 0
          }
        }
      },
      "ErrorBody": {
        "type": "object",
        "description": "Why a request failed.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
      "Folder": {
        "type": "object",
        "description": "One top-level folder in /projects.",
        "required": [
          "name"
        ],
        "properties": {
          "git": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/GitDetails",
                "description": "Present when the folder is a git repository."
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "type": "string",
            "description": "The folder's name in /projects."
          }
        }
      },
      "FolderStatus": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Folder"
          },
          {
            "type": "object",
            "required": [
              "serve"
            ],
            "properties": {
              "serve": {
                "type": "boolean",
                "description": "Chosen to be served to the Claude app, by its own choice or by default."
              }
            }
          }
        ],
        "description": "A folder with what the Claude app sees of it."
      },
      "GitDetails": {
        "type": "object",
        "description": "Where a repository comes from and what it has checked out.",
        "properties": {
          "branch": {
            "type": [
              "string",
              "null"
            ],
            "description": "The checked-out branch, absent when not on a branch."
          },
          "repository": {
            "type": [
              "string",
              "null"
            ],
            "description": "Where `origin` points, without credentials."
          }
        }
      },
...
      "LoginPrompt": {
        "type": "object",
        "description": "What the person signing in needs: a page to open, and sometimes a code to enter there.",
        "required": [
          "url"
        ],
        "properties": {
          "code": {
            "type": [
              "string",
              "null"
            ],
            "description": "The one-time code to enter on that page, when the sign-in uses one."
          },
          "url": {
            "type": "string",
            "description": "The sign-in page to open."
          }
        }
      },
...
      "ServeBody": {
        "type": "object",
        "description": "Whether the Claude app lists a folder.",
        "required": [
          "serve"
        ],
        "properties": {
          "serve": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}